Crate rustfmt_config_proc_macro

This crate provides a derive macro for ConfigType.

Attribute Macros

config_type

nightly_only_test

Used to conditionally output the TokenStream for tests that need to be run on nightly only.

rustfmt_only_ci_test

Used to conditionally output the TokenStream for tests that should be run as part of rustfmts test suite, but should be ignored when running in the rust-lang/rust test suite.

stable_only_test

Used to conditionally output the TokenStream for tests that need to be run on stable only.